pyproject.toml
[project]
name = "django-business-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.2.3",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "markdown>=3.8",
]
